requirements.txt
pytesseract==0.2.6
Pillow==6.0.0
numpy==1.16.2
opencv-contrib-python==4.0.0.21
